Restructuring CC example.
* Added non-Markovian system dynamics.
* Added MC validation step.
* Moved all CC work to new directory.
