+++ examples/chance_constrained/cc_experiment_matrix.py
@@ -1,7 +1,13 @@
 import numpy as np
 
-from examples.benchmark_chance_constrained_problem import ex as cc_experiment
-from examples.benchmark_chance_constrained_problem_pd import ex as cc_experiment_pd
+from examples.chance_constrained.benchmark_chance_constrained_problem import (
+    ex as cc_experiment,
+)
+from examples.chance_constrained.benchmark_chance_constrained_problem_pd import (
+    ex as cc_experiment_pd,
+)
+
+from examples.chance_constrained.cc_env import NonMarkovIntegratorEnv
 
 
 def ex_matrix():
@@ -105,5 +111,7 @@ def ex_pd_vs_non():
 
 
 if __name__ == "__main__":
-    ex_matrix()
+    # ex_matrix()
     # ex_pd_vs_non()
+
+    test_system()
