+++ examples/chance_constrained/cc_problem.py
@@ -19,18 +19,11 @@ Example:
         IEEE Conference on Decision and Control,
         <https://arxiv.org/abs/2103.12759>`_
 
-Todo:
-    * Matrix of kernel parameter vs. sample size.
-    * Update to Mixed paper obstacles.
-    * Validate against mixed paper soln.
-
 """
 
 
 import gym
 import gym_socks
-
-import matplotlib.pyplot as plt
 
 import logging
 
@@ -38,22 +31,18 @@ import numpy as np
 from numpy.linalg import norm
 
 from sacred import Experiment
-from sacred import Ingredient
 
 from functools import partial
-from sklearn.metrics.pairwise import rbf_kernel, euclidean_distances
-from gym_socks.kernel.metrics import abel_kernel
+from sklearn.metrics.pairwise import rbf_kernel
 
-from gym_socks.algorithms.control import KernelControlCC
+from gym_socks.algorithms.control import KernelControlFwd
+from gym_socks.algorithms.control import KernelControlBwd
 
 from gym_socks.envs.sample import sample as _sample, transpose_sample
-
-# from gym_socks.envs.sample import trajectory_sampler
-from gym_socks.envs.sample import sample_generator
 from gym_socks.envs.sample import trajectory_sampler
 from gym_socks.envs.sample import reshape_trajectory_sample
 
-from gym_socks.envs.policy import RandomizedPolicy, BasePolicy, PDController
+from gym_socks.envs.policy import RandomizedPolicy, BasePolicy
 
 from gym_socks.envs.dynamical_system import DynamicalSystem
 
@@ -78,7 +67,7 @@ from examples.ingredients.plotting_ingredient import update_rc_params
 @system_ingredient.config
 def system_config():
 
-    system_id = "RepeatedIntegratorEnv-v0"
+    system_id = "NonMarkovIntegratorEnv-v0"
 
     sampling_time = 0.1
 
@@ -109,64 +98,7 @@ def sample_config():
 @simulation_ingredient.config
 def simulation_config():
 
-    initial_condition = [-0.5, 0.1, -0.5, 0.1]
-
-
-cost_ingredient = Ingredient("cost")
-
-
-@cost_ingredient.config
-def cost_config():
-
-    goal = [0.5, 0, 0.5, 0]
-
-
-@cost_ingredient.capture
-def make_cost(time_horizon, goal):
-
-    goal = np.array(goal)
-
-    def _cost_fn(time, state):
-        state = np.reshape(state, (-1, time_horizon, 4))
-        dist = state[:, -1, [0, 2]] - np.array([goal[[0, 2]]])
-        result = np.linalg.norm(dist, ord=2, axis=1)
-        return result
-        # return np.power(result, 2)
-
-    return _cost_fn
-
-
-obstacle_ingredient = Ingredient("obstacle")
-
-
-@obstacle_ingredient.config
-def obstacle_config():
-
-    center = [0, 0]
-    radius = 0.2
-
-
-constraint_ingredient = Ingredient("constraint", ingredients=[obstacle_ingredient])
-
-
-@constraint_ingredient.capture
-def make_constraint(time_horizon, obstacle):
-
-    center = np.array(obstacle["center"])
-
-    def _constraint_fn(time, state):
-        state = np.reshape(state, (-1, time_horizon, 4))
-        dist = state[:, :, [0, 2]] - np.array([obstacle["center"]])
-        result = np.linalg.norm(dist, ord=2, axis=2)
-        indicator = np.all(result >= obstacle["radius"], axis=1)
-        return indicator
-        # return 0.2 - np.power(result, 2) - 1 + delta
-        # dist >= 0.2
-        # 0 >= 0.2 - dist
-
-        # E[f(X, U)] >= 1 - delta
-
-    return _constraint_fn
+    initial_condition = [-0.5, 0, -0.5, 0]
 
 
 ex = Experiment(
@@ -175,9 +107,7 @@ ex = Experiment(
         simulation_ingredient,
         sample_ingredient,
         plotting_ingredient,
-        cost_ingredient,
-        constraint_ingredient,
-    ]
+    ],
 )
 
 
@@ -200,26 +130,25 @@ def config(sample):
 
         To specify configuration variables, use `with variable=value`, e.g.
 
-            $ python -m <experiment> with seed=123 time_horizon=5
+            $ python -m <experiment> with seed=123 system.time_horizon=5
 
     .. _sacred:
         https://sacred.readthedocs.io/en/stable/index.html
 
     """
 
-    sigma = 5  # Kernel bandwidth parameter.
+    sigma = 3  # Kernel bandwidth parameter.
     # Regularization parameter.
-    regularization_param = 1e-7
+    regularization_param = 1e-5
 
     # Probability of violation.
-    delta = 0.01
+    delta = 0.1
 
-    time_horizon = 15
+    time_horizon = 25
 
     verbose = True
 
     gen_sample = True
-    pd_gains = [[3, 0.5, 0, 0], [0, 0, 3, 0.5]]
 
     results_filename = "results/data.npy"
     no_plot = False
@@ -234,8 +163,6 @@ def main(
     time_horizon,
     verbose,
     gen_sample,
-    cost,
-    pd_gains,
     results_filename,
     no_plot,
     sample,
@@ -261,29 +188,12 @@ def main(
         )
         sample_space.seed(seed=seed)
 
-        # TODO define goal state and pass through
-        PD_gains = np.array(pd_gains)
-        # ---------------------------------------------------------
-        # DEBUGGING CODE
-        # dt = 0.1
-        # A = np.array([[1,dt,0,0],[0,1,0,0],[0,0,1,dt],[0,0,0,1]])
-        # B = np.array([[dt**2/2,0],[dt,0],[0,dt**2/2],[0,dt]])
-        # np.linalg.eig(A+B@K)[0]
-        # ---------------------------------------------------------
-        ClosedLoopPDPolicy = PDController(
-            action_space=env.action_space,
-            state_space=env.state_space,
-            goal_state=np.array(cost["goal"]),
-            PD_gains=PD_gains,
-        )
         S = _sample(
             sampler=trajectory_sampler(
                 time_horizon=time_horizon,
                 env=env,
-                policy=ClosedLoopPDPolicy,
+                policy=RandomizedPolicy(action_space=env.action_space),
                 sample_space=sample_space,
-                # init_state = simulation["initial_condition"],
-                # B_initialize_from_init_state=True
             ),
             sample_size=sample["sample_space"]["sample_size"],
         )
@@ -306,15 +216,9 @@ def main(
         #    u  - np.array (N*u_dim) # vector (N is time_horizon)
         #    x  - np.array (N*x_dim) # vector (N is time_horizon)
 
-        # -------------------------------------------------
-        # DEBUG
         X, U, Y = transpose_sample(S)
-        # trajs = np.array(Y)
-        # plt.figure(0)
-        # for i in range(trajs.shape[0]):
-        #     plt.plot(trajs[i, :, 0], trajs[i, :, 2], alpha=0.2)
-        # plt.show()
-        # -------------------------------------------------
+        # print(np.array(X[0]))
+        # print(np.array(Y[0]))
 
         # Generate the set of admissible control actions.
         _log.debug("Generating admissible control actions.")
@@ -322,7 +226,7 @@ def main(
             sampler=trajectory_sampler(
                 time_horizon=time_horizon,
                 env=env,
-                policy=ClosedLoopPDPolicy,
+                policy=RandomizedPolicy(action_space=env.action_space),
                 sample_space=sample_space,
             ),
             sample_size=_get_sample_size(env.action_space, sample["action_space"]),
@@ -346,21 +250,37 @@ def main(
     S = list(zip(X, U, Y))
 
     T = reshape_trajectory_sample(S)
-    # T = normalize_trajectory_sample(T)
     # Define the cost and constraint functions.
+
+    def _cost_fn(time, state):
+        state = np.reshape(state, (-1, time_horizon, 4))
+        # print(np.shape(state))
+        # print(np.shape(state[:, -1, [0, 2]]))
+        dist = state[:, -1, [0, 2]] - np.array([[0.5, 0.5]])
+        result = np.linalg.norm(dist, ord=2, axis=1)
+        return np.power(result, 2)
+
+    def _constraint_fn(time, state):
+        state = np.reshape(state, (-1, time_horizon, 4))
+        # print(state[0][0])
+        dist = state[:, :, [0, 2]] - np.array([[0, 0]])
+        result = np.linalg.norm(dist, ord=2, axis=2)
+        indicator = np.any(np.power(result, 2) <= 0.2, axis=1)
+        return 1 - delta - indicator
+        # return 0.2 - np.power(result, 2) - 1 + delta
+        # dist >= 0.2
+        # 0 >= 0.2 - dist
+
+        # E[f(X, U)] >= 1 - delta
 
     with ComputationTimer():
 
         # Compute policy.
-        policy = KernelControlCC(
-            cost_fn=make_cost(time_horizon=time_horizon),
-            constraint_fn=make_constraint(time_horizon=time_horizon),
-            delta=delta,
+        policy = KernelControlFwd(
+            cost_fn=_cost_fn,
+            constraint_fn=_constraint_fn,
             verbose=verbose,
             kernel_fn=partial(rbf_kernel, gamma=1 / (2 * (sigma ** 2))),
-            # kernel_fn=partial(
-            #     abel_kernel, sigma=sigma, distance_fn=euclidean_distances
-            # ),
             regularization_param=regularization_param,
             seed=seed,
         )
@@ -391,10 +311,10 @@ def main(
 
 @plotting_ingredient.config_hook
 def plot_config(config, command_name, logger):
-    if command_name in {"main", "plot_results", "plot_sample"}:
+    if command_name in {"main", "plot_results"}:
         return {
             "trajectory_style": {
-                "lines.linewidth": 0.5,
+                "lines.linewidth": 1,
                 "lines.linestyle": "-",
                 "lines.color": "C1",
             },
@@ -408,7 +328,10 @@ def plot_config(config, command_name, logger):
 
 
 @ex.command(unobserved=True)
-def plot_results(system, obstacle, plot_cfg):
+def plot_results(
+    system,
+    plot_cfg,
+):
     """Plot the results of the experiement."""
 
     logging.getLogger("matplotlib").setLevel(logging.WARNING)
@@ -437,62 +360,9 @@ def plot_results(system, obstacle, plot_cfg):
             label="System Trajectory",
         )
 
-    plt.gca().add_patch(
-        plt.Circle(
-            obstacle["center"],
-            obstacle["radius"],
-            fc="none",
-            ec="red",
-            label="Obstacle",
-        )
-    )
+    plt.gca().add_patch(plt.Circle((0, 0), 0.2, fc="none", ec="red"))
 
     plt.legend()
-
-    plt.savefig(plot_cfg["plot_filename"])
-
-
-@ex.command(unobserved=True)
-def plot_sample(time_horizon, plot_cfg):
-
-    logging.getLogger("matplotlib").setLevel(logging.WARNING)
-    logging.getLogger("PIL").setLevel(logging.WARNING)
-
-    # Dynamically load for speed.
-    import matplotlib
-
-    matplotlib.use("Agg")
-    update_rc_params(matplotlib, plot_cfg["rc_params"])
-
-    import matplotlib.pyplot as plt
-
-    with open("results/sample.npy", "rb") as f:
-        X = np.load(f).tolist()
-        U = np.load(f).tolist()
-        Y = np.load(f).tolist()
-        A = np.load(f).tolist()
-
-    constraint_fn = make_constraint(time_horizon=time_horizon)
-    satisfies_constraints = constraint_fn(time=0, state=Y)
-
-    fig = plt.figure()
-    ax = plt.axes(**plot_cfg["axes"])
-
-    for i, trajectory in enumerate(Y):
-
-        trajectory = np.array(trajectory, dtype=np.float32)
-
-        if satisfies_constraints[i] == True:
-            plt_color = "C0"
-        else:
-            plt_color = "C1"
-
-        with plt.style.context(plot_cfg["trajectory_style"]):
-            plt.plot(trajectory[:, 0], trajectory[:, 2], color=plt_color, marker="")
-
-    plt.gca().add_patch(plt.Circle((0, 0), 0.2, fc="none", ec="red", label="Obstacle"))
-
-    plt.scatter(0.5, 0.5, s=2.5, c="C2")
 
     plt.savefig(plot_cfg["plot_filename"])
